requirements.txt
Pillow==10.0.0
PyMuPDF==1.23.3
PyMuPDFb==1.23.3
